added register page

+++ src/components/LoginForm.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, Link } from "react-router-dom";
 
 import "../css/LoginForm.css";
 import { RiErrorWarningFill } from "react-icons/ri";
@@ -8,17 +8,19 @@ import Swal from "sweetalert2";
 import withReactContent from "sweetalert2-react-content";
 
 // firebase
-import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth, provider } from "../components/firebase";
+import {
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+} from "firebase/auth";
+import { auth } from "../components/firebase";
 
 const MySwal = withReactContent(Swal);
 
-const LoginForm = ({ updateState }) => {
+const LoginForm = ({ updateState, formType }) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [isPassword, setIsPassword] = useState(true);
   const [isEmail, setIsEmail] = useState(true);
-  // const [isValid, setIsValid] = useState(true);
   let navigate = useNavigate();
 
   // Regex for email validation
@@ -31,35 +33,52 @@ const LoginForm = ({ updateState }) => {
 
   const handleSubmit = (e) => {
     e.preventDefault();
-    // if (email.length === 0 || password.length === 0) {
-    //   setIsValid(!isValid);
-    // } else {
-    //   setIsValid(true);
-    // }
+
     // sign-in
     if (isEmail && isPassword) {
-      console.log("pass");
-      signInWithEmailAndPassword(auth, email, password)
-        .then((userCredential) => {
-          // Signed in
-          MySwal.fire({
-            title: "Logged In!",
-            icon: "success",
+      if (formType === "Register") {
+        createUserWithEmailAndPassword(auth, email, password)
+          .then((userCredential) => {
+            // Signed in
+            MySwal.fire({
+              title: "Account created",
+              icon: "success",
+            });
+            localStorage.setItem("user", JSON.stringify(userCredential.user));
+            setEmail("");
+            setPassword("");
+            updateState();
+            navigate("/student/dashboard");
+          })
+          .catch((error) => {
+            const errorCode = error.code.slice(5).replace(/-/g, " ");
+            MySwal.fire({
+              title: errorCode,
+              icon: "warning",
+            });
           });
-          // console.log(userCredential.user);
-          localStorage.setItem("user", JSON.stringify(userCredential.user));
-          setEmail("");
-          setPassword("");
-          updateState();
-          navigate("/student/dashboard");
-        })
-        .catch((error) => {
-          const errorCode = error.code.slice(5).replace(/-/g, " ");
-          MySwal.fire({
-            title: errorCode,
-            icon: "warning",
+      } else {
+        signInWithEmailAndPassword(auth, email, password)
+          .then((userCredential) => {
+            // Signed in
+            MySwal.fire({
+              title: "Logged In!",
+              icon: "success",
+            });
+            localStorage.setItem("user", JSON.stringify(userCredential.user));
+            setEmail("");
+            setPassword("");
+            updateState();
+            navigate("/student/dashboard");
+          })
+          .catch((error) => {
+            const errorCode = error.code.slice(5).replace(/-/g, " ");
+            MySwal.fire({
+              title: errorCode,
+              icon: "warning",
+            });
           });
-        });
+      }
     }
   };
   return (
@@ -74,7 +93,7 @@ const LoginForm = ({ updateState }) => {
           value={email}
           onChange={(e) => {
             setEmail(e.target.value);
-            if (!e.target.value.match(pattern) && e.target.value.length != 0) {
+            if (!e.target.value.match(pattern) && e.target.value.length !== 0) {
               setIsEmail(false);
             } else {
               setIsEmail(true);
@@ -106,9 +125,16 @@ const LoginForm = ({ updateState }) => {
         {!isPassword && <RiErrorWarningFill className="warning" />}
       </label>
       {/* <p className="err-message">{err}</p> */}
-      <button type="submit" onClick={handleSubmit}>
-        Login
-      </button>
+      <div className="cta">
+        {formType !== "Register" && (
+          <Link to="/register" className="create-account">
+            Create account
+          </Link>
+        )}
+        <button type="submit" onClick={handleSubmit}>
+          {formType === "Register" ? "Sign up" : "Login"}
+        </button>
+      </div>
     </form>
   );
 };
